[https://nvbugs/6403920][fix] Move allreduce strategy test to smoke, use full-warp hidden_size (#16389)
